some fixes, add lambda func, add eip for api ec2
i didn't run terraform plan/apply in the prev commit so i never knew
there were some errors. i added lambda.tf for the lambda function that
resets the leaderboard bi-weekly, and code for it (boilerplate for now).
also added an elastic ip address for the web server ec2 so that the api
can still be accessed even after the instance restarts