add eventbridge rule for lambda, change provider version
this will be used for triggering the lambda function daily. the idea is
to trigger it daily and then check in the code whether the date of the
invocation matches one of the dates in the database or something, where
i'll be keeping track of dates when staff sets up new climbs at the
basement wall.